rename `cross_validate_item.py` to `cross_validation_item.py`
