+++ skore/src/skore/ui/project_routes.py
@@ -9,7 +9,7 @@ from fastapi import APIRouter, HTTPException, Request, status
 from fastapi.params import Depends
 from fastapi.templating import Jinja2Templates
 
-from skore.item.cross_validate_item import CrossValidationItem
+from skore.item.cross_validation_item import CrossValidationItem
 from skore.item.media_item import MediaItem
 from skore.item.numpy_array_item import NumpyArrayItem
 from skore.item.pandas_dataframe_item import PandasDataFrameItem
